Move EmbeddedMetricsMessage
