+++ src/recorder/serialization/embedded_metrics_message.h
