+++ test/recorder/stream_recorder/BUILD
@@ -76,16 +76,6 @@ lightstep_catch_test(
 )
 
 lightstep_catch_test(
-    name = "embedded_metrics_message_test",
-    srcs = [
-        "embedded_metrics_message_test.cpp",
-    ],
-    deps = [
-        "//src/recorder/stream_recorder:embedded_metrics_message_lib",
-    ],
-)
-
-lightstep_catch_test(
     name = "span_stream_test",
     srcs = [
         "span_stream_test.cpp",
